requirements.txt
fastapi
uvicorn[standard]
pillow
tensorflow
numpy
python-multipart
pathlib
